requirements.txt
Django==2.1.3
djangorestframework==3.9.0
feedparser==5.2.1
pytz==2018.7
